feat(tools): add new tools for media management and scraping
- Each tool includes functionalities for scanning directories, translating NFO files, testing crawlers, cleaning up files, finding missing numbers, syncing person data, scraping single files, and managing symlinks.
- Implemented user-friendly interfaces with input fields, buttons, and feedback messages to improve user experience.
- Integrated IPC calls for backend communication, ensuring smooth operation of each tool.
- Added necessary utility functions and styles to support the new components, enhancing overall application consistency and maintainability.
