+++ src/renderer/src/components/tool/toolStyles.ts
@@ -1,0 +1,17 @@
+export const TOOL_PANEL_CLASS =
+  "rounded-quiet-lg border-none bg-surface-floating/96 py-0 shadow-[0_18px_44px_rgba(15,23,42,0.05)]";
+export const TOOL_SUBSECTION_CLASS = "space-y-4 rounded-quiet-lg bg-surface-low/90 p-4 md:p-5";
+export const TOOL_INPUT_CLASS =
+  "h-11 rounded-quiet-sm border-none bg-surface-low/90 px-4 shadow-none focus-visible:ring-2 focus-visible:ring-ring/30";
+export const TOOL_TEXTAREA_CLASS =
+  "rounded-quiet-sm border-none bg-surface-low/90 p-4 shadow-none focus-visible:ring-2 focus-visible:ring-ring/30";
+export const TOOL_SELECT_TRIGGER_CLASS =
+  "h-11 rounded-quiet-sm border-none bg-surface-low/90 px-4 shadow-none focus-visible:ring-2 focus-visible:ring-ring/30";
+export const TOOL_PRIMARY_BUTTON_CLASS =
+  "h-11 rounded-quiet-capsule bg-primary px-5 text-sm font-semibold text-primary-foreground hover:bg-primary/90";
+export const TOOL_SECONDARY_BUTTON_CLASS =
+  "h-11 rounded-quiet-capsule bg-surface-low px-5 text-sm font-semibold text-foreground hover:bg-surface-raised/75";
+export const TOOL_ICON_BUTTON_CLASS =
+  "h-11 w-11 shrink-0 rounded-quiet-sm bg-surface-low text-foreground hover:bg-surface-raised/75";
+export const TOOL_TABLE_SHELL_CLASS = "overflow-hidden rounded-quiet-lg bg-surface-floating/96";
+export const TOOL_NOTE_CLASS = "text-xs leading-6 text-muted-foreground";
