+++ src/renderer/src/routes/tool.tsx
@@ -1,2056 +1,59 @@
-import type { Website } from "@shared/enums";
-import { toErrorMessage } from "@shared/error";
-import type {
-  AmazonPosterScanItem,
-  BatchTranslateApplyResultItem,
-  BatchTranslateScanItem,
-  EmbyConnectionCheckResult,
-  JellyfinConnectionCheckResult,
-  PersonSyncResult,
-} from "@shared/ipcTypes";
-import { useMutation, useQuery } from "@tanstack/react-query";
-import { createFileRoute, useNavigate } from "@tanstack/react-router";
-import {
-  ChevronLeft,
-  ChevronRight,
-  ClipboardList,
-  Eraser,
-  FileCheck,
-  FileSearch,
-  FolderOpen,
-  Link2,
-  Search,
-  Settings2,
-  ShoppingCart,
-  Trash2,
-  UserCheck,
-  Wrench,
-} from "lucide-react";
-import { type MutableRefObject, useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { deleteFile } from "@/api/manual";
-import { createSymlink, listEntries, scrapeSingleFile } from "@/client/api";
-import { ipc } from "@/client/ipc";
-import { chooseScrapeFilePath } from "@/client/scrapeFilePath";
-import type { CreateSoftlinksBody, FileItem, ScrapeFileBody } from "@/client/types";
-import { AmazonPosterDialog } from "@/components/AmazonPosterDialog";
-import { PageHeader } from "@/components/PageHeader";
-import { Badge } from "@/components/ui/Badge";
+import { createFileRoute } from "@tanstack/react-router";
+import { ArrowLeft } from "lucide-react";
+import { useRef, useState } from "react";
+import { ToolDetail } from "@/components/tool/ToolDetail";
+import { ToolOverview } from "@/components/tool/ToolOverview";
+import type { ToolId } from "@/components/tool/toolCatalog";
 import { Button } from "@/components/ui/Button";
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/Card";
-import { Checkbox } from "@/components/ui/Checkbox";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/Dialog";
-import { Input } from "@/components/ui/Input";
-import { Label } from "@/components/ui/Label";
-import { Progress } from "@/components/ui/Progress";
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/Select";
-import { TabButton } from "@/components/ui/TabButton";
-import { Textarea } from "@/components/ui/Textarea";
-import { useToast } from "@/contexts/ToastProvider";
-import { cn } from "@/lib/utils";
-import { useUIStore } from "@/store/uiStore";
-import { formatBytes } from "@/utils/format";
 
 export const Route = createFileRoute("/tool")({
   component: ToolComponent,
 });
 
-interface MissingResultRow {
-  index: number;
-  number: string;
-}
-
-interface CleanupCandidate {
-  path: string;
-  name: string;
-  ext: string;
-  size: number;
-  lastModified: string | null;
-}
-
-type SyncMode = "all" | "missing";
-type ConnectionCheckResult = JellyfinConnectionCheckResult | EmbyConnectionCheckResult;
-
-const CLEANUP_PRESET_EXTENSIONS = [".html", ".url", ".txt", ".nfo", ".jpg", ".png", ".torrent", ".ass", ".srt"];
-const CLEANUP_MAX_SCANNED_DIRECTORIES = 50000;
-
-const clearProgressResetTimer = (timerRef: MutableRefObject<number | null>) => {
-  if (timerRef.current !== null) {
-    window.clearTimeout(timerRef.current);
-    timerRef.current = null;
-  }
-};
-
-const TOOLS_TABS = [
-  { id: "scraping", label: "数据刮削", icon: FileCheck },
-  { id: "maintenance", label: "维护管理", icon: Settings2 },
-  { id: "cleanup", label: "垃圾清理", icon: Eraser },
-  { id: "utility", label: "实用工具", icon: ClipboardList },
-] as const;
-
-function toVisitedDirectoryKey(dirPath: string) {
-  const trimmed = dirPath.trim();
-  if (!trimmed) return "";
-
-  const withoutTrailingSeparators = trimmed.replace(/[\\/]+$/u, "");
-  return (withoutTrailingSeparators || trimmed).toLowerCase();
-}
-
-function escapeRegExp(text: string) {
-  return text.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-
-function parseRangeInput(text: string): { start: number; end: number; width: number } | null {
-  const match = text.trim().match(/^(\d+)\s*[-~—到]+\s*(\d+)$/u);
-  if (!match) return null;
-  const startRaw = match[1];
-  const endRaw = match[2];
-  const start = Number(startRaw);
-  const end = Number(endRaw);
-  if (!Number.isInteger(start) || !Number.isInteger(end) || start < 0 || end < 0 || start > end) {
-    return null;
-  }
-  return { start, end, width: Math.max(startRaw.length, endRaw.length, 3) };
-}
-
-function normalizeExtension(ext: string) {
-  const val = ext.trim().toLowerCase();
-  if (!val) return "";
-  return val.startsWith(".") ? val : `.${val}`;
-}
-
-function extensionFromName(fileName: string) {
-  const dot = fileName.lastIndexOf(".");
-  if (dot < 0) return "";
-  return normalizeExtension(fileName.slice(dot));
-}
-
-function getFirstDiagnosticError(result: ConnectionCheckResult) {
-  return result.steps.find((step) => step.status === "error");
-}
-
-function getFirstDiagnosticBlocker(result: ConnectionCheckResult) {
-  return getFirstDiagnosticError(result) ?? result.steps.find((step) => step.status !== "ok");
-}
-
-function canRunPersonSync(result: ConnectionCheckResult | null): result is ConnectionCheckResult {
-  return Boolean(result?.success);
-}
-
-function getDiagnosticHeadline(result: ConnectionCheckResult) {
-  if (!result.success) {
-    return "存在阻塞项";
-  }
-  if (result.personCount === 0) {
-    return "人物库为空";
-  }
-  return "可以执行人物同步";
-}
-
-function getEmptyPersonLibraryMessage(serverName: "Jellyfin" | "Emby", targetLabel: "人物信息" | "人物头像") {
-  return `${serverName} 人物库为空。已确认连接与权限状态正常，当前无法执行${targetLabel}同步。请先在 ${serverName} 中生成人物条目后重试。`;
-}
-
-function formatSyncResult(label: string, result: PersonSyncResult) {
-  return `${label}: 成功 ${result.processedCount}，失败 ${result.failedCount}，跳过 ${result.skippedCount}`;
-}
-
-interface PersonToolCardProps {
-  title: string;
-  description: string;
-  diagnosticLabel: string;
-  checkResult: ConnectionCheckResult | null;
-  checkPending: boolean;
-  syncState: {
-    busy: boolean;
-    infoSyncRunning: boolean;
-    photoSyncRunning: boolean;
-    progress: number;
-    infoMode: SyncMode;
-    photoMode: SyncMode;
-  };
-  texts: {
-    infoMissing: string;
-    infoAll: string;
-    photoMissing: string;
-    photoAll: string;
-    photoNotice?: string;
-  };
-  actions: {
-    onCheck: () => void;
-    onInfoModeChange: (value: SyncMode) => void;
-    onPhotoModeChange: (value: SyncMode) => void;
-    onSyncInfo: () => void;
-    onSyncPhoto: () => void;
-  };
-  headerExtra?: React.ReactNode;
-  className?: string;
-}
-
-function PersonToolCard({
-  title,
-  description,
-  diagnosticLabel,
-  checkResult,
-  checkPending,
-  syncState,
-  texts,
-  actions,
-  headerExtra,
-  className,
-}: PersonToolCardProps) {
-  const { busy, infoSyncRunning, photoSyncRunning, progress, infoMode, photoMode } = syncState;
-  const { infoMissing, infoAll, photoMissing, photoAll, photoNotice } = texts;
-  const { onCheck, onInfoModeChange, onPhotoModeChange, onSyncInfo, onSyncPhoto } = actions;
-
-  return (
-    <Card className={cn("rounded-xl border shadow-sm", className)}>
-      <CardHeader>
-        <div className="flex flex-wrap items-start justify-between gap-3">
-          <div className="flex min-w-0 items-center gap-2">
-            <div className="p-1.5 bg-primary/8 rounded-lg">
-              <UserCheck className="h-5 w-5 text-primary" />
-            </div>
-            <div className="min-w-0">
-              <CardTitle className="text-sm font-medium">{title}</CardTitle>
-              <CardDescription className="text-xs">{description}</CardDescription>
-            </div>
-          </div>
-          <div className="flex shrink-0 flex-wrap items-center gap-2">
-            {headerExtra}
-            <Button
-              variant="outline"
-              onClick={onCheck}
-              disabled={checkPending || busy}
-              className="rounded-lg shrink-0 h-9 text-sm"
-            >
-              {checkPending ? "诊断中..." : "连接诊断"}
-            </Button>
-          </div>
-        </div>
-      </CardHeader>
-      <CardContent className="space-y-6">
-        {checkResult && (
-          <div className="rounded-xl border bg-muted/20 p-4 space-y-3">
-            <div className="flex items-center justify-between gap-3">
-              <div>
-                <div className="text-xs font-medium text-muted-foreground">{diagnosticLabel}</div>
-                {(checkResult.serverInfo?.serverName || checkResult.serverInfo?.version) && (
-                  <div className="text-sm">
-                    {[checkResult.serverInfo?.serverName, checkResult.serverInfo?.version].filter(Boolean).join(" ")}
-                  </div>
-                )}
-              </div>
-              <div
-                className={cn(
-                  "text-xs font-medium",
-                  !checkResult.success && "text-amber-600",
-                  checkResult.success && checkResult.personCount === 0 && "text-muted-foreground",
-                  checkResult.success && checkResult.personCount !== 0 && "text-emerald-600",
-                )}
-              >
-                {getDiagnosticHeadline(checkResult)}
-              </div>
-            </div>
-            <div className="grid gap-2">
-              {checkResult.steps.map((step) => (
-                <div
-                  key={step.key}
-                  className="flex items-start justify-between gap-3 rounded-lg bg-background/70 px-3 py-2"
-                >
-                  <div>
-                    <div className="text-xs font-medium">{step.label}</div>
-                    <div className="text-xs text-muted-foreground">{step.message}</div>
-                  </div>
-                  <div
-                    className={cn(
-                      "text-[11px] font-medium shrink-0",
-                      step.status === "ok" && "text-emerald-600",
-                      step.status === "error" && "text-red-600",
-                      step.status === "skipped" && "text-muted-foreground",
-                    )}
-                  >
-                    {step.status === "ok" ? "通过" : step.status === "error" ? "失败" : "跳过"}
-                  </div>
-                </div>
-              ))}
-            </div>
-          </div>
-        )}
-
-        <div className="space-y-4">
-          <div className="grid gap-2">
-            <Label className="text-xs font-medium text-muted-foreground">演员资料同步</Label>
-            <div className="flex gap-2">
-              <Select value={infoMode} onValueChange={(value) => onInfoModeChange(value as SyncMode)}>
-                <SelectTrigger className="h-9 bg-muted/30 rounded-lg border-none focus:ring-2 flex-1">
-                  <SelectValue />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="missing">仅补全空白资料</SelectItem>
-                  <SelectItem value="all">更新已有资料</SelectItem>
-                </SelectContent>
-              </Select>
-              <Button
-                variant="secondary"
-                onClick={onSyncInfo}
-                disabled={busy || checkPending}
-                className="flex-1 rounded-lg h-9 text-sm"
-              >
-                {infoSyncRunning ? "同步中..." : "同步信息"}
-              </Button>
-            </div>
-            <div className="text-[11px] leading-relaxed text-muted-foreground">
-              {infoMode === "missing" ? infoMissing : infoAll}
-            </div>
-          </div>
-
-          <div className="grid gap-2">
-            <Label className="text-xs font-medium text-muted-foreground">演员头像同步</Label>
-            <div className="flex gap-2">
-              <Select value={photoMode} onValueChange={(value) => onPhotoModeChange(value as SyncMode)}>
-                <SelectTrigger className="h-9 bg-muted/30 rounded-lg border-none focus:ring-2 flex-1">
-                  <SelectValue />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="missing">仅补全缺失头像</SelectItem>
-                  <SelectItem value="all">重新同步头像</SelectItem>
-                </SelectContent>
-              </Select>
-              <Button
-                variant="secondary"
-                onClick={onSyncPhoto}
-                disabled={busy || checkPending}
-                className="flex-1 rounded-lg h-9 text-sm"
-              >
-                {photoSyncRunning ? "同步中..." : "同步头像"}
-              </Button>
-            </div>
-            <div className="text-[11px] leading-relaxed text-muted-foreground">
-              {photoMode === "missing" ? photoMissing : photoAll}
-            </div>
-            {photoNotice && <div className="text-[11px] leading-relaxed text-amber-700">{photoNotice}</div>}
-          </div>
-
-          {progress > 0 && (
-            <div className="grid gap-2">
-              <div className="flex justify-between text-xs text-muted-foreground font-medium">
-                <span>任务进度</span>
-                <span>{Math.round(progress)}%</span>
-              </div>
-              <Progress value={progress} className="h-2 bg-muted/30" />
-            </div>
-          )}
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
-
 function ToolComponent() {
-  const navigate = useNavigate();
-  const { showSuccess, showError, showInfo } = useToast();
-  const toolActiveTab = useUIStore((state) => state.toolActiveTab);
-  const setToolActiveTab = useUIStore((state) => state.setToolActiveTab);
-  const activeTab: (typeof TOOLS_TABS)[number]["id"] = TOOLS_TABS.some((tab) => tab.id === toolActiveTab)
-    ? (toolActiveTab as (typeof TOOLS_TABS)[number]["id"])
-    : "scraping";
+  const pageScrollRef = useRef<HTMLDivElement>(null);
+  const [selectedToolId, setSelectedToolId] = useState<ToolId | null>(null);
 
-  // 单文件刮削
-  const [singleFilePath, setSingleFilePath] = useState("");
-  const scrapeSingleFileMut = useMutation({
-    mutationFn: async (body: ScrapeFileBody) => scrapeSingleFile({ body, throwOnError: true }),
-  });
-
-  const [sourceDir, setSourceDir] = useState("");
-  const [destDir, setDestDir] = useState("");
-  const [copyFiles, setCopyFiles] = useState(false);
-  const createSymlinkMut = useMutation({
-    mutationFn: async (body: CreateSoftlinksBody) => createSymlink({ body, throwOnError: true }),
-  });
-
-  // 演员相关
-  const checkJellyfinConnectionMut = useMutation({
-    mutationFn: async () => ipc.tool.checkJellyfinConnection(),
-  });
-  const checkEmbyConnectionMut = useMutation({
-    mutationFn: async () => ipc.tool.checkEmbyConnection(),
-  });
-  const [jellyfinCheckResult, setJellyfinCheckResult] = useState<JellyfinConnectionCheckResult | null>(null);
-  const [embyCheckResult, setEmbyCheckResult] = useState<EmbyConnectionCheckResult | null>(null);
-  const [jellyfinActorInfoMode, setJellyfinActorInfoMode] = useState<SyncMode>("missing");
-  const [jellyfinActorPhotoMode, setJellyfinActorPhotoMode] = useState<SyncMode>("missing");
-  const [embyActorInfoMode, setEmbyActorInfoMode] = useState<SyncMode>("missing");
-  const [embyActorPhotoMode, setEmbyActorPhotoMode] = useState<SyncMode>("missing");
-  const [jellyfinInfoSyncRunning, setJellyfinInfoSyncRunning] = useState(false);
-  const [jellyfinPhotoSyncRunning, setJellyfinPhotoSyncRunning] = useState(false);
-  const [embyInfoSyncRunning, setEmbyInfoSyncRunning] = useState(false);
-  const [embyPhotoSyncRunning, setEmbyPhotoSyncRunning] = useState(false);
-  const [jellyfinSyncProgress, setJellyfinSyncProgress] = useState(0);
-  const [embySyncProgress, setEmbySyncProgress] = useState(0);
-  const jellyfinProgressResetTimerRef = useRef<number | null>(null);
-  const embyProgressResetTimerRef = useRef<number | null>(null);
-  const jellyfinSyncRunning = jellyfinInfoSyncRunning || jellyfinPhotoSyncRunning;
-  const embySyncRunning = embyInfoSyncRunning || embyPhotoSyncRunning;
-  const anyPersonSyncRunning = jellyfinSyncRunning || embySyncRunning;
-  const anyPersonCheckPending = checkJellyfinConnectionMut.isPending || checkEmbyConnectionMut.isPending;
-  const [selectedPersonServer, setSelectedPersonServer] = useState<"jellyfin" | "emby">("jellyfin");
-
-  // 缺番查找
-  const [missingPrefix, setMissingPrefix] = useState("");
-  const [missingRange, setMissingRange] = useState("");
-  const [existingNumbers, setExistingNumbers] = useState("");
-  const [missingRows, setMissingRows] = useState<MissingResultRow[]>([]);
-  const [missingSummary, setMissingSummary] = useState("");
-
-  // 文件清理
-  const [cleanPath, setCleanPath] = useState("");
-  const [cleanExtensions, setCleanExtensions] = useState<string[]>([".html", ".url"]);
-  const [cleanCustomExt, setCleanCustomExt] = useState("");
-  const [includeSubdirs, setIncludeSubdirs] = useState(true);
-  const [cleanupScanning, setCleanupScanning] = useState(false);
-  const [cleanupDeleting, setCleanupDeleting] = useState(false);
-  const [cleanupProgress, setCleanupProgress] = useState(0);
-  const [cleanupConfirmOpen, setCleanupConfirmOpen] = useState(false);
-  const [cleanupCandidates, setCleanupCandidates] = useState<CleanupCandidate[]>([]);
-
-  // 爬虫测试
-  const sitesQ = useQuery({
-    queryKey: ["crawler", "sites"],
-    queryFn: async () => {
-      const result = (await ipc.crawler.listSites()) as {
-        sites: Array<{ site: string; name: string; enabled: boolean; native: boolean }>;
-      };
-      return result.sites;
-    },
-  });
-  const [crawlerTestSite, setCrawlerTestSite] = useState("");
-  const [crawlerTestNumber, setCrawlerTestNumber] = useState("");
-  const [crawlerTestResult, setCrawlerTestResult] = useState<{
-    data: {
-      title?: string;
-      actors?: string[];
-      genres?: string[];
-      release_date?: string;
-      studio?: string;
-    } | null;
-    error?: string;
-    elapsed: number;
-  } | null>(null);
-  const [crawlerTesting, setCrawlerTesting] = useState(false);
-  const [amazonDir, setAmazonDir] = useState("");
-  const [amazonPosterDialogOpen, setAmazonPosterDialogOpen] = useState(false);
-  const [amazonPosterScanItems, setAmazonPosterScanItems] = useState<AmazonPosterScanItem[]>([]);
-  const [amazonScanning, setAmazonScanning] = useState(false);
-  const [batchTranslateDir, setBatchTranslateDir] = useState("");
-  const [batchTranslateItems, setBatchTranslateItems] = useState<BatchTranslateScanItem[]>([]);
-  const [batchTranslateResults, setBatchTranslateResults] = useState<BatchTranslateApplyResultItem[]>([]);
-  const [batchTranslateScanning, setBatchTranslateScanning] = useState(false);
-  const [batchTranslateApplying, setBatchTranslateApplying] = useState(false);
-
-  // Navigation arrows logic
-  const scrollRef = useRef<HTMLDivElement>(null);
-  const [canScrollLeft, setCanScrollLeft] = useState(false);
-  const [canScrollRight, setCanScrollRight] = useState(false);
-
-  const checkScroll = useCallback(() => {
-    if (scrollRef.current) {
-      const { scrollLeft, scrollWidth, clientWidth } = scrollRef.current;
-      setCanScrollLeft(scrollLeft > 2);
-      setCanScrollRight(scrollLeft < scrollWidth - clientWidth - 2);
-    }
-  }, []);
-
-  useEffect(() => {
-    checkScroll();
-    window.addEventListener("resize", checkScroll);
-    return () => window.removeEventListener("resize", checkScroll);
-  }, [checkScroll]);
-
-  const scroll = (direction: "left" | "right") => {
-    if (scrollRef.current) {
-      const { clientWidth } = scrollRef.current;
-      const scrollAmount = clientWidth * 0.75;
-      scrollRef.current.scrollBy({
-        left: direction === "left" ? -scrollAmount : scrollAmount,
-        behavior: "smooth",
-      });
-    }
-  };
-
-  useEffect(() => {
-    const el = scrollRef.current;
-    if (!el) return;
-
-    const onWheelNative = (e: WheelEvent) => {
-      if (el.scrollWidth > el.clientWidth) {
-        if (Math.abs(e.deltaY) > Math.abs(e.deltaX)) {
-          e.preventDefault();
-          e.stopPropagation();
-          el.scrollLeft += e.deltaY;
-        } else if (Math.abs(e.deltaX) > 0) {
-          e.stopPropagation();
-        }
-      }
-    };
-
-    el.addEventListener("wheel", onWheelNative, { passive: false });
-    return () => el.removeEventListener("wheel", onWheelNative);
-  }, []);
-
-  useEffect(() => {
-    return ipc.on.progress((payload) => {
-      if (jellyfinSyncRunning) {
-        setJellyfinSyncProgress(payload.value);
-        return;
-      }
-      if (embySyncRunning) {
-        setEmbySyncProgress(payload.value);
-      }
+  const scrollToTop = () => {
+    window.requestAnimationFrame(() => {
+      pageScrollRef.current?.scrollTo({ top: 0, behavior: "smooth" });
     });
-  }, [embySyncRunning, jellyfinSyncRunning]);
-
-  useEffect(() => {
-    return () => {
-      clearProgressResetTimer(jellyfinProgressResetTimerRef);
-      clearProgressResetTimer(embyProgressResetTimerRef);
-    };
-  }, []);
-
-  const handleScrapeSingleFile = async () => {
-    if (!singleFilePath) {
-      showError("请输入文件路径");
-      return;
-    }
-    showInfo("正在启动单文件刮削任务...");
-    try {
-      const result = await scrapeSingleFileMut.mutateAsync({
-        path: singleFilePath,
-      });
-      showSuccess(result.data.message);
-      setTimeout(() => navigate({ to: "/logs" }), 1000);
-    } catch (error) {
-      showError(`单文件刮削任务启动失败: ${error}`);
-    }
   };
 
-  const handleBrowseSingleFile = async () => {
-    try {
-      const selectedPath = await chooseScrapeFilePath();
-      if (selectedPath) {
-        setSingleFilePath(selectedPath);
-      }
-    } catch (error) {
-      showError(`文件选择失败: ${toErrorMessage(error)}`);
-    }
+  const handleSelectTool = (toolId: ToolId) => {
+    setSelectedToolId(toolId);
+    scrollToTop();
   };
 
-  const handleCreateSymlink = async () => {
-    if (!sourceDir || !destDir) {
-      showError("请输入源目录和目标目录");
-      return;
-    }
-
-    showInfo("正在启动软链接创建任务...");
-    try {
-      const result = await createSymlinkMut.mutateAsync({
-        source_dir: sourceDir,
-        dest_dir: destDir,
-        copy_files: copyFiles,
-      });
-      showSuccess(result.data.message);
-      setTimeout(() => navigate({ to: "/logs" }), 1000);
-    } catch (error) {
-      showError(`软链接创建任务启动失败: ${toErrorMessage(error)}`);
-    }
+  const handleBackToOverview = () => {
+    setSelectedToolId(null);
+    scrollToTop();
   };
-
-  const runJellyfinConnectionCheck = async (silentSuccess = false): Promise<JellyfinConnectionCheckResult | null> => {
-    try {
-      const result = await checkJellyfinConnectionMut.mutateAsync();
-      setJellyfinCheckResult(result);
-
-      const firstError = getFirstDiagnosticError(result);
-      if (!firstError) {
-        if (!silentSuccess) {
-          showSuccess("Jellyfin 连接诊断通过");
-        }
-      } else if (!silentSuccess) {
-        showError(`${firstError.label}: ${firstError.message}`);
-      }
-      return result;
-    } catch (error) {
-      showError(`Jellyfin 连通性测试失败: ${toErrorMessage(error)}`);
-      setJellyfinCheckResult(null);
-      return null;
-    }
-  };
-
-  const runEmbyConnectionCheck = async (silentSuccess = false): Promise<EmbyConnectionCheckResult | null> => {
-    try {
-      const result = await checkEmbyConnectionMut.mutateAsync();
-      setEmbyCheckResult(result);
-
-      const firstError = getFirstDiagnosticError(result);
-      if (!firstError) {
-        if (!silentSuccess) {
-          showSuccess("Emby 连接诊断通过");
-        }
-      } else if (!silentSuccess) {
-        showError(`${firstError.label}: ${firstError.message}`);
-      }
-      return result;
-    } catch (error) {
-      showError(`Emby 连通性测试失败: ${toErrorMessage(error)}`);
-      setEmbyCheckResult(null);
-      return null;
-    }
-  };
-
-  const handleCheckJellyfinConnection = async () => {
-    showInfo("正在诊断 Jellyfin 连接状态...");
-    await runJellyfinConnectionCheck();
-  };
-
-  const handleCheckEmbyConnection = async () => {
-    showInfo("正在诊断 Emby 连接状态...");
-    await runEmbyConnectionCheck();
-  };
-
-  const handleSyncJellyfinActorInfo = async () => {
-    showInfo("正在诊断 Jellyfin 连接状态...");
-    const diagnostic = await runJellyfinConnectionCheck(true);
-    if (!canRunPersonSync(diagnostic)) {
-      const blocker = diagnostic ? getFirstDiagnosticBlocker(diagnostic) : undefined;
-      if (blocker) {
-        showError(`${blocker.label}: ${blocker.message}`);
-      }
-      return;
-    }
-    if (diagnostic.personCount === 0) {
-      showInfo(getEmptyPersonLibraryMessage("Jellyfin", "人物信息"));
-      return;
-    }
-
-    clearProgressResetTimer(jellyfinProgressResetTimerRef);
-    setJellyfinSyncProgress(0);
-    setJellyfinInfoSyncRunning(true);
-    showInfo("正在同步 Jellyfin 演员信息...");
-    try {
-      const result = await ipc.tool.syncJellyfinActorInfo(jellyfinActorInfoMode);
-      setJellyfinSyncProgress(100);
-      showSuccess(formatSyncResult("Jellyfin 演员信息同步完成", result));
-    } catch (error) {
-      showError(`Jellyfin 演员信息同步失败: ${toErrorMessage(error)}`);
-    } finally {
-      setJellyfinInfoSyncRunning(false);
-      clearProgressResetTimer(jellyfinProgressResetTimerRef);
-      jellyfinProgressResetTimerRef.current = window.setTimeout(() => {
-        setJellyfinSyncProgress(0);
-        jellyfinProgressResetTimerRef.current = null;
-      }, 1200);
-    }
-  };
-
-  const handleSyncJellyfinPhotos = async () => {
-    showInfo("正在诊断 Jellyfin 连接状态...");
-    const diagnostic = await runJellyfinConnectionCheck(true);
-    if (!canRunPersonSync(diagnostic)) {
-      const blocker = diagnostic ? getFirstDiagnosticBlocker(diagnostic) : undefined;
-      if (blocker) {
-        showError(`${blocker.label}: ${blocker.message}`);
-      }
-      return;
-    }
-    if (diagnostic.personCount === 0) {
-      showInfo(getEmptyPersonLibraryMessage("Jellyfin", "人物头像"));
-      return;
-    }
-
-    clearProgressResetTimer(jellyfinProgressResetTimerRef);
-    setJellyfinSyncProgress(0);
-    setJellyfinPhotoSyncRunning(true);
-    showInfo("正在同步 Jellyfin 演员头像...");
-    try {
-      const result = await ipc.tool.syncJellyfinActorPhoto(jellyfinActorPhotoMode);
-      setJellyfinSyncProgress(100);
-      showSuccess(formatSyncResult("Jellyfin 头像同步完成", result));
-    } catch (error) {
-      showError(`Jellyfin 头像同步失败: ${toErrorMessage(error)}`);
-    } finally {
-      setJellyfinPhotoSyncRunning(false);
-      clearProgressResetTimer(jellyfinProgressResetTimerRef);
-      jellyfinProgressResetTimerRef.current = window.setTimeout(() => {
-        setJellyfinSyncProgress(0);
-        jellyfinProgressResetTimerRef.current = null;
-      }, 1200);
-    }
-  };
-
-  const handleSyncEmbyActorInfo = async () => {
-    showInfo("正在诊断 Emby 连接状态...");
-    const diagnostic = await runEmbyConnectionCheck(true);
-    if (!canRunPersonSync(diagnostic)) {
-      const blocker = diagnostic ? getFirstDiagnosticBlocker(diagnostic) : undefined;
-      if (blocker) {
-        showError(`${blocker.label}: ${blocker.message}`);
-      }
-      return;
-    }
-    if (diagnostic.personCount === 0) {
-      showInfo(getEmptyPersonLibraryMessage("Emby", "人物信息"));
-      return;
-    }
-
-    clearProgressResetTimer(embyProgressResetTimerRef);
-    setEmbySyncProgress(0);
-    setEmbyInfoSyncRunning(true);
-    showInfo("正在同步 Emby 演员信息...");
-    try {
-      const result = await ipc.tool.syncEmbyActorInfo(embyActorInfoMode);
-      setEmbySyncProgress(100);
-      showSuccess(formatSyncResult("Emby 演员信息同步完成", result));
-    } catch (error) {
-      showError(`Emby 演员信息同步失败: ${toErrorMessage(error)}`);
-    } finally {
-      setEmbyInfoSyncRunning(false);
-      clearProgressResetTimer(embyProgressResetTimerRef);
-      embyProgressResetTimerRef.current = window.setTimeout(() => {
-        setEmbySyncProgress(0);
-        embyProgressResetTimerRef.current = null;
-      }, 1200);
-    }
-  };
-
-  const handleSyncEmbyPhotos = async () => {
-    showInfo("正在诊断 Emby 连接状态...");
-    const diagnostic = await runEmbyConnectionCheck(true);
-    if (!canRunPersonSync(diagnostic)) {
-      const blocker = diagnostic ? getFirstDiagnosticBlocker(diagnostic) : undefined;
-      if (blocker) {
-        showError(`${blocker.label}: ${blocker.message}`);
-      }
-      return;
-    }
-    if (diagnostic.personCount === 0) {
-      showInfo(getEmptyPersonLibraryMessage("Emby", "人物头像"));
-      return;
-    }
-
-    const adminKeyStep = diagnostic.steps.find((step) => step.key === "adminKey");
-    if (adminKeyStep?.message) {
-      showInfo(adminKeyStep.message);
-    }
-
-    clearProgressResetTimer(embyProgressResetTimerRef);
-    setEmbySyncProgress(0);
-    setEmbyPhotoSyncRunning(true);
-    showInfo("正在同步 Emby 演员头像...");
-    try {
-      const result = await ipc.tool.syncEmbyActorPhoto(embyActorPhotoMode);
-      setEmbySyncProgress(100);
-      showSuccess(formatSyncResult("Emby 头像同步完成", result));
-    } catch (error) {
-      showError(`Emby 头像同步失败: ${toErrorMessage(error)}`);
-    } finally {
-      setEmbyPhotoSyncRunning(false);
-      clearProgressResetTimer(embyProgressResetTimerRef);
-      embyProgressResetTimerRef.current = window.setTimeout(() => {
-        setEmbySyncProgress(0);
-        embyProgressResetTimerRef.current = null;
-      }, 1200);
-    }
-  };
-
-  const toggleCleanExtension = (extension: string) => {
-    const normalized = normalizeExtension(extension);
-    if (!normalized) return;
-    setCleanExtensions((prev) =>
-      prev.includes(normalized) ? prev.filter((ext) => ext !== normalized) : [...prev, normalized],
-    );
-  };
-
-  const handleAddCustomExtension = () => {
-    const normalized = normalizeExtension(cleanCustomExt);
-    if (!normalized) {
-      showError("请输入有效的扩展名");
-      return;
-    }
-    if (cleanExtensions.includes(normalized)) {
-      setCleanCustomExt("");
-      showInfo(`文件类型 ${normalized} 已存在`);
-      return;
-    }
-    setCleanExtensions((prev) => [...prev, normalized]);
-    setCleanCustomExt("");
-  };
-
-  const browseDirectory = async (onSelect: (path: string) => void) => {
-    const result = await ipc.file.browse("directory");
-    const selectedPath = result.paths?.[0];
-    if (selectedPath) {
-      onSelect(selectedPath);
-    }
-  };
-
-  const scanCleanupCandidates = async () => {
-    const targetPath = cleanPath.trim();
-    if (!targetPath) {
-      showError("请输入需要扫描的目录");
-      return;
-    }
-    if (cleanExtensions.length === 0) {
-      showError("请至少选择一种文件类型");
-      return;
-    }
-
-    setCleanupScanning(true);
-    setCleanupCandidates([]);
-    setCleanupProgress(0);
-
-    const extensionSet = new Set(cleanExtensions.map(normalizeExtension).filter(Boolean));
-    const found: CleanupCandidate[] = [];
-    const queue: string[] = [targetPath];
-    const visited = new Set<string>();
-
-    try {
-      while (queue.length > 0) {
-        const current = queue.shift();
-        if (!current) continue;
-
-        const currentKey = toVisitedDirectoryKey(current);
-        if (!currentKey || visited.has(currentKey)) continue;
-        visited.add(currentKey);
-        if (visited.size > CLEANUP_MAX_SCANNED_DIRECTORIES) {
-          throw new Error(`扫描目录数量超过 ${CLEANUP_MAX_SCANNED_DIRECTORIES}，请缩小路径范围后重试`);
-        }
-
-        const response = await listEntries({ query: { path: current }, throwOnError: true });
-        const items = response.data?.items ?? [];
-        for (const item of items) {
-          if (item.type === "directory") {
-            if (includeSubdirs) queue.push(item.path);
-            continue;
-          }
-          if (!shouldKeepForCleanup(item, extensionSet)) continue;
-          found.push({
-            path: item.path,
-            name: item.name,
-            ext: extensionFromName(item.name),
-            size: item.size ?? 0,
-            lastModified: item.last_modified ?? null,
-          });
-        }
-      }
-
-      found.sort((a, b) => a.path.localeCompare(b.path, "zh-CN"));
-      setCleanupCandidates(found);
-      if (found.length === 0) {
-        showInfo("未找到匹配文件。");
-      } else {
-        showSuccess(`扫描完成，共找到 ${found.length} 个匹配文件。`);
-      }
-    } catch (error) {
-      showError(`扫描失败: ${toErrorMessage(error)}`);
-    } finally {
-      setCleanupScanning(false);
-    }
-  };
-
-  const handleDeleteCleanupCandidates = async () => {
-    if (cleanupCandidates.length === 0) {
-      showInfo("当前没有可清理文件。");
-      return;
-    }
-
-    setCleanupDeleting(true);
-    setCleanupProgress(0);
-
-    const failedPaths = new Set<string>();
-    let successCount = 0;
-
-    try {
-      for (const [index, candidate] of cleanupCandidates.entries()) {
-        try {
-          await deleteFile(candidate.path);
-          successCount += 1;
-        } catch {
-          failedPaths.add(candidate.path);
-        }
-        setCleanupProgress(Math.round(((index + 1) / cleanupCandidates.length) * 100));
-      }
-      setCleanupCandidates((prev) => prev.filter((item) => failedPaths.has(item.path)));
-      setCleanupConfirmOpen(false);
-      if (failedPaths.size === 0) {
-        showSuccess(`文件清理完成，成功删除 ${successCount} 个文件。`);
-      } else {
-        showError(`删除完成：成功 ${successCount}，失败 ${failedPaths.size}。`);
-      }
-    } finally {
-      setCleanupDeleting(false);
-      window.setTimeout(() => setCleanupProgress(0), 1200);
-    }
-  };
-
-  const handleFindMissing = () => {
-    const prefix = missingPrefix.trim().toUpperCase();
-    if (!prefix) {
-      showError("请输入番号前缀");
-      return;
-    }
-
-    const range = parseRangeInput(missingRange);
-    if (!range) {
-      showError("请输入有效范围，例如 1-200 或 001-120");
-      return;
-    }
-    if (range.end - range.start > 20000) {
-      showError("范围过大，请缩小后再查询");
-      return;
-    }
-
-    const matchedNumbers = new Set<number>();
-    const re = new RegExp(`${escapeRegExp(prefix)}[-_\\s]?(\\d+)`, "giu");
-    for (const match of existingNumbers.toUpperCase().matchAll(re)) {
-      const num = Number(match[1]);
-      if (Number.isInteger(num)) matchedNumbers.add(num);
-    }
-    if (matchedNumbers.size === 0) {
-      for (const raw of existingNumbers.match(/\d+/g) ?? []) {
-        const num = Number(raw);
-        if (Number.isInteger(num)) matchedNumbers.add(num);
-      }
-    }
-
-    const rows: MissingResultRow[] = [];
-    for (let i = range.start; i <= range.end; i += 1) {
-      if (!matchedNumbers.has(i)) {
-        rows.push({
-          index: rows.length + 1,
-          number: `${prefix}-${String(i).padStart(range.width, "0")}`,
-        });
-      }
-    }
-
-    const expectedTotal = range.end - range.start + 1;
-    setMissingRows(rows);
-    setMissingSummary(
-      `范围 ${range.start}-${range.end}，期望 ${expectedTotal}，已识别 ${matchedNumbers.size}，缺失 ${rows.length}`,
-    );
-    if (rows.length === 0) {
-      showSuccess("未发现缺失番号");
-    } else {
-      showInfo(`查找完成，缺失 ${rows.length} 条`);
-    }
-  };
-
-  const handleCrawlerTest = async () => {
-    if (!crawlerTestSite) {
-      showError("请选择站点");
-      return;
-    }
-    if (!crawlerTestNumber.trim()) {
-      showError("请输入番号");
-      return;
-    }
-    setCrawlerTesting(true);
-    setCrawlerTestResult(null);
-    try {
-      const result = await ipc.crawler.test(crawlerTestSite as Website, crawlerTestNumber.trim());
-      setCrawlerTestResult(result);
-      if (result.data) {
-        showSuccess(`测试成功，耗时 ${(result.elapsed / 1000).toFixed(1)}s`);
-      } else {
-        showError(result.error ?? "未获取到数据");
-      }
-    } catch (error) {
-      showError(`爬虫测试失败: ${toErrorMessage(error)}`);
-    } finally {
-      setCrawlerTesting(false);
-    }
-  };
-
-  const handleAmazonPosterScan = async () => {
-    const directory = amazonDir.trim();
-    if (!directory) {
-      showError("请输入需要扫描的媒体目录");
-      return;
-    }
-
-    setAmazonScanning(true);
-    try {
-      const result = await ipc.tool.amazonPosterScan(directory);
-      setAmazonPosterScanItems(result.items);
-      setAmazonPosterDialogOpen(true);
-
-      if (result.items.length === 0) {
-        showInfo("扫描完成，但未找到可处理的 NFO 条目。");
-      } else {
-        showSuccess(`扫描完成，共找到 ${result.items.length} 个条目。`);
-      }
-    } catch (error) {
-      showError(`Amazon 海报扫描失败: ${toErrorMessage(error)}`);
-    } finally {
-      setAmazonScanning(false);
-    }
-  };
-
-  const scanBatchTranslateItems = async (options: { silent?: boolean } = {}) => {
-    const directory = batchTranslateDir.trim();
-    if (!directory) {
-      setBatchTranslateItems([]);
-      showError("请输入需要扫描的媒体目录");
-      return null;
-    }
-
-    setBatchTranslateScanning(true);
-    setBatchTranslateItems([]);
-    try {
-      const result = await ipc.tool.batchTranslateScan(directory);
-      setBatchTranslateItems(result.items);
-
-      if (!options.silent) {
-        if (result.items.length === 0) {
-          showInfo("扫描完成，未发现待翻译的 NFO 条目。");
-        } else {
-          const fieldCount = result.items.reduce((sum, item) => sum + item.pendingFields.length, 0);
-          showSuccess(`扫描完成，共找到 ${result.items.length} 个条目，待处理字段 ${fieldCount} 项。`);
-        }
-      }
-
-      return result.items;
-    } catch (error) {
-      setBatchTranslateItems([]);
-      showError(`批量翻译扫描失败: ${toErrorMessage(error)}`);
-      return null;
-    } finally {
-      setBatchTranslateScanning(false);
-    }
-  };
-
-  const handleBatchTranslateScan = async () => {
-    setBatchTranslateResults([]);
-    await scanBatchTranslateItems();
-  };
-
-  const handleBatchTranslateApply = async () => {
-    if (batchTranslateItems.length === 0) {
-      showInfo("当前没有待翻译条目。");
-      return;
-    }
-
-    setBatchTranslateApplying(true);
-    try {
-      const result = await ipc.tool.batchTranslateApply(batchTranslateItems);
-      setBatchTranslateResults(result.results);
-
-      const successCount = result.results.filter((item) => item.success).length;
-      const partialCount = result.results.filter((item) => !item.success && item.translatedFields.length > 0).length;
-      const failedCount = result.results.length - successCount - partialCount;
-
-      if (failedCount === 0) {
-        showSuccess(`批量翻译完成：成功 ${successCount}，部分成功 ${partialCount}。`);
-      } else {
-        showError(`批量翻译完成：成功 ${successCount}，部分成功 ${partialCount}，失败 ${failedCount}。`);
-      }
-
-      await scanBatchTranslateItems({ silent: true });
-    } catch (error) {
-      showError(`批量翻译执行失败: ${toErrorMessage(error)}`);
-    } finally {
-      setBatchTranslateApplying(false);
-    }
-  };
-
-  const cleanupTotalSize = useMemo(
-    () => cleanupCandidates.reduce((sum, item) => sum + (Number.isFinite(item.size) ? item.size : 0), 0),
-    [cleanupCandidates],
-  );
-  const cleanupPreviewRows = cleanupCandidates.slice(0, 400);
-  const missingPreviewRows = missingRows.slice(0, 300);
-  const batchTranslatePreviewRows = batchTranslateItems.slice(0, 300);
-  const batchTranslateResultRows = batchTranslateResults.slice(0, 300);
-  const batchTranslatePendingFieldCount = useMemo(
-    () => batchTranslateItems.reduce((sum, item) => sum + item.pendingFields.length, 0),
-    [batchTranslateItems],
-  );
-
-  const personToolProps =
-    selectedPersonServer === "jellyfin"
-      ? {
-          diagnosticLabel: "Jellyfin 诊断结果",
-          checkResult: jellyfinCheckResult,
-          checkPending: checkJellyfinConnectionMut.isPending,
-          syncState: {
-            busy: anyPersonSyncRunning,
-            infoSyncRunning: jellyfinInfoSyncRunning,
-            photoSyncRunning: jellyfinPhotoSyncRunning,
-            progress: jellyfinSyncProgress,
-            infoMode: jellyfinActorInfoMode,
-            photoMode: jellyfinActorPhotoMode,
-          },
-          texts: {
-            infoMissing: "仅补全缺失的演员简介与基础资料。",
-            infoAll: "按当前抓取结果更新演员简介与基础资料。",
-            photoMissing: "仅为缺少头像的演员补充头像。",
-            photoAll: "按当前抓取结果重新同步演员头像。",
-            photoNotice: undefined as string | undefined,
-          },
-          actions: {
-            onCheck: handleCheckJellyfinConnection,
-            onInfoModeChange: setJellyfinActorInfoMode,
-            onPhotoModeChange: setJellyfinActorPhotoMode,
-            onSyncInfo: handleSyncJellyfinActorInfo,
-            onSyncPhoto: handleSyncJellyfinPhotos,
-          },
-        }
-      : {
-          diagnosticLabel: "Emby 诊断结果",
-          checkResult: embyCheckResult,
-          checkPending: checkEmbyConnectionMut.isPending,
-          syncState: {
-            busy: anyPersonSyncRunning,
-            infoSyncRunning: embyInfoSyncRunning,
-            photoSyncRunning: embyPhotoSyncRunning,
-            progress: embySyncProgress,
-            infoMode: embyActorInfoMode,
-            photoMode: embyActorPhotoMode,
-          },
-          texts: {
-            infoMissing: "仅补全缺失的演员简介与基础资料，并保留未变更字段。",
-            infoAll: "按当前抓取结果更新演员简介与基础资料，并按同步字段写回 Emby。",
-            photoMissing: "仅为缺少头像的演员补充头像。",
-            photoAll: "按当前抓取结果重新同步演员头像。",
-            photoNotice: "人物头像上传通常需要管理员 API Key。若返回 401 或 403，请改用管理员 API Key 后重试。",
-          },
-          actions: {
-            onCheck: handleCheckEmbyConnection,
-            onInfoModeChange: setEmbyActorInfoMode,
-            onPhotoModeChange: setEmbyActorPhotoMode,
-            onSyncInfo: handleSyncEmbyActorInfo,
-            onSyncPhoto: handleSyncEmbyPhotos,
-          },
-        };
 
   return (
-    <div className="h-full w-full overflow-y-auto relative scroll-smooth">
-      <div className="sticky top-0 z-10 bg-background/60 backdrop-blur-xl border-b">
-        <PageHeader title="工具" subtitle="常用批量任务与维护工具" icon={Wrench} />
-
-        <div className="px-8 pb-2 h-11 flex items-center">
-          <div className="relative flex items-center flex-1 min-w-0 group/tabs">
-            {canScrollLeft && (
-              <div className="absolute left-0 inset-y-0 z-10 flex items-center pr-10 bg-gradient-to-r from-background via-background/80 to-transparent pointer-events-none rounded-l-lg">
-                <Button
-                  type="button"
-                  variant="ghost"
-                  size="icon"
-                  className="h-7 w-7 ml-0.5 rounded-full bg-background shadow-md border pointer-events-auto hover:bg-accent hover:text-accent-foreground transition-all duration-200"
-                  onClick={() => scroll("left")}
-                >
-                  <ChevronLeft className="h-4 w-4" />
-                </Button>
-              </div>
-            )}
-
-            <div
-              ref={scrollRef}
-              onScroll={checkScroll}
-              className="flex-1 flex gap-1 p-1 bg-muted/40 rounded-lg overflow-x-auto no-scrollbar scroll-smooth"
-            >
-              {TOOLS_TABS.map((tab) => {
-                const Icon = tab.icon;
-                return (
-                  <TabButton key={tab.id} isActive={activeTab === tab.id} onClick={() => setToolActiveTab(tab.id)}>
-                    <Icon className="h-3.5 w-3.5 mr-1.5" />
-                    {tab.label}
-                  </TabButton>
-                );
-              })}
-            </div>
-
-            {canScrollRight && (
-              <div className="absolute right-0 inset-y-0 z-10 flex items-center pl-10 bg-gradient-to-l from-background via-background/80 to-transparent pointer-events-none rounded-r-lg">
-                <Button
-                  type="button"
-                  variant="ghost"
-                  size="icon"
-                  className="h-7 w-7 mr-0.5 rounded-full bg-background shadow-md border pointer-events-auto hover:bg-accent hover:text-accent-foreground transition-all duration-200"
-                  onClick={() => scroll("right")}
-                >
-                  <ChevronRight className="h-4 w-4" />
-                </Button>
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-
-      <div className="max-w-5xl mx-auto p-6 space-y-8">
-        {/* ── 数据刮削 ─────────────────────────── */}
-        {activeTab === "scraping" && (
-          <div className="space-y-8 animate-in fade-in slide-in-from-bottom-4 duration-500">
-            <div className="px-1">
-              <h2 className="text-base font-semibold mb-1">数据刮削</h2>
-              <p className="text-muted-foreground text-xs">按需处理单个文件，并测试爬虫抓取能力</p>
-            </div>
-
-            <div className="grid gap-6 md:grid-cols-1">
-              <Card className="rounded-xl border shadow-sm flex flex-col">
-                <CardHeader className="pb-3">
-                  <div className="flex items-center gap-2">
-                    <div className="p-1.5 bg-primary/8 rounded-lg">
-                      <FileSearch className="h-5 w-5 text-primary" />
-                    </div>
-                    <div>
-                      <CardTitle className="text-sm font-medium">单文件刮削</CardTitle>
-                      <CardDescription className="text-xs">输入文件路径，快速处理指定视频</CardDescription>
-                    </div>
-                  </div>
-                </CardHeader>
-                <CardContent className="space-y-4">
-                  <div className="grid gap-2">
-                    <Label htmlFor="filePath" className="text-xs font-medium text-muted-foreground">
-                      文件路径
-                    </Label>
-                    <div className="flex items-center gap-2">
-                      <Input
-                        id="filePath"
-                        value={singleFilePath}
-                        onChange={(e) => setSingleFilePath(e.target.value)}
-                        placeholder="/path/to/video.mp4"
-                        className="h-9 bg-muted/30 rounded-lg border-none focus:ring-2"
-                      />
-                      <Button
-                        type="button"
-                        variant="outline"
-                        onClick={handleBrowseSingleFile}
-                        className="h-9 shrink-0 rounded-lg px-3"
-                      >
-                        <FolderOpen className="h-4 w-4" />
-                      </Button>
-                    </div>
-                  </div>
-                  <Button
-                    variant="secondary"
-                    onClick={handleScrapeSingleFile}
-                    disabled={scrapeSingleFileMut.isPending}
-                    className="w-full rounded-lg h-9 text-sm font-medium"
-                  >
-                    {scrapeSingleFileMut.isPending ? "正在刮削..." : "开始单文件刮削"}
-                  </Button>
-                </CardContent>
-              </Card>
-            </div>
-
-            {/* 爬虫测试 */}
-            <Card className="rounded-xl border shadow-sm">
-              <CardHeader className="pb-3">
-                <div className="flex items-center gap-2">
-                  <div className="p-1.5 bg-primary/8 rounded-lg">
-                    <Search className="h-5 w-5 text-primary" />
-                  </div>
-                  <div>
-                    <CardTitle className="text-sm font-medium">爬虫测试</CardTitle>
-                    <CardDescription className="text-xs">选择站点和番号，测试爬虫是否能正确抓取数据</CardDescription>
-                  </div>
-                </div>
-              </CardHeader>
-              <CardContent className="space-y-4">
-                <div className="grid gap-4 md:grid-cols-2">
-                  <div className="grid gap-2">
-                    <Label className="text-xs font-medium text-muted-foreground">站点</Label>
-                    <Select value={crawlerTestSite} onValueChange={setCrawlerTestSite}>
-                      <SelectTrigger className="h-9 bg-muted/30 rounded-lg border-none focus:ring-2">
-                        <SelectValue placeholder="选择站点" />
-                      </SelectTrigger>
-                      <SelectContent>
-                        {(sitesQ.data ?? []).map((s) => (
-                          <SelectItem key={s.site} value={s.site}>
-                            <span className="flex items-center gap-2">
-                              {s.name}
-                              {s.enabled && (
-                                <Badge variant="secondary" className="h-4 text-[9px] px-1">
-                                  已启用
-                                </Badge>
-                              )}
-                              {!s.native && (
-                                <Badge variant="outline" className="h-4 text-[9px] px-1">
-                                  浏览器
-                                </Badge>
-                              )}
-                            </span>
-                          </SelectItem>
-                        ))}
-                      </SelectContent>
-                    </Select>
-                  </div>
-                  <div className="grid gap-2">
-                    <Label htmlFor="crawlerTestNumber" className="text-xs font-medium text-muted-foreground">
-                      番号
-                    </Label>
-                    <Input
-                      id="crawlerTestNumber"
-                      value={crawlerTestNumber}
-                      onChange={(e) => setCrawlerTestNumber(e.target.value)}
-                      placeholder="例如: ABP-001"
-                      className="h-9 bg-muted/30 rounded-lg border-none focus:ring-2"
-                      onKeyDown={(e) => {
-                        if (e.key === "Enter") handleCrawlerTest();
-                      }}
-                    />
-                  </div>
-                </div>
-                <Button
-                  variant="secondary"
-                  onClick={handleCrawlerTest}
-                  disabled={crawlerTesting}
-                  className="w-full rounded-lg h-9 text-sm font-medium"
-                >
-                  {crawlerTesting ? "测试中..." : "开始测试"}
-                </Button>
-
-                {crawlerTestResult && (
-                  <div className="rounded-xl border bg-muted/10 p-4 space-y-2">
-                    <div className="flex items-center justify-between text-xs">
-                      <span className="font-medium">
-                        {crawlerTestResult.data ? (
-                          <span className="text-green-600">测试成功</span>
-                        ) : (
-                          <span className="text-destructive">测试失败</span>
-                        )}
-                      </span>
-                      <span className="text-muted-foreground">
-                        耗时 {(crawlerTestResult.elapsed / 1000).toFixed(1)}s
-                      </span>
-                    </div>
-                    {crawlerTestResult.error && <p className="text-xs text-destructive">{crawlerTestResult.error}</p>}
-                    {crawlerTestResult.data && (
-                      <div className="space-y-1 text-xs">
-                        {crawlerTestResult.data.title && (
-                          <div>
-                            <span className="text-muted-foreground">标题: </span>
-                            <span className="font-medium">{crawlerTestResult.data.title}</span>
-                          </div>
-                        )}
-                        {crawlerTestResult.data.actors && crawlerTestResult.data.actors.length > 0 && (
-                          <div>
-                            <span className="text-muted-foreground">演员: </span>
-                            <span>{crawlerTestResult.data.actors.join(", ")}</span>
-                          </div>
-                        )}
-                        {crawlerTestResult.data.genres && crawlerTestResult.data.genres.length > 0 && (
-                          <div>
-                            <span className="text-muted-foreground">标签: </span>
-                            <span>{crawlerTestResult.data.genres.join(", ")}</span>
-                          </div>
-                        )}
-                        {crawlerTestResult.data.release_date && (
-                          <div>
-                            <span className="text-muted-foreground">发行日期: </span>
-                            <span>{crawlerTestResult.data.release_date}</span>
-                          </div>
-                        )}
-                        {crawlerTestResult.data.studio && (
-                          <div>
-                            <span className="text-muted-foreground">片商: </span>
-                            <span>{crawlerTestResult.data.studio}</span>
-                          </div>
-                        )}
-                      </div>
-                    )}
-                  </div>
-                )}
-              </CardContent>
-            </Card>
-
-            <Card className="rounded-xl border shadow-sm">
-              <CardHeader className="pb-3">
-                <div className="flex items-center gap-2">
-                  <div className="p-1.5 bg-primary/8 rounded-lg">
-                    <ShoppingCart className="h-5 w-5 text-primary" />
-                  </div>
-                  <div>
-                    <CardTitle className="text-sm font-medium">Amazon 海报增强</CardTitle>
-                    <CardDescription className="text-xs">从 Amazon.co.jp 查询高质量竖版海报图片</CardDescription>
-                  </div>
-                </div>
-              </CardHeader>
-              <CardContent className="space-y-4">
-                <div className="grid gap-2">
-                  <Label htmlFor="amazon-poster-dir" className="text-xs font-medium text-muted-foreground">
-                    目标目录
-                  </Label>
-                  <div className="flex gap-2">
-                    <Input
-                      id="amazon-poster-dir"
-                      value={amazonDir}
-                      onChange={(e) => setAmazonDir(e.target.value)}
-                      placeholder="输入已刮削完成的输出目录"
-                      className="h-9 bg-muted/30 rounded-lg border-none focus:ring-2 flex-1"
-                    />
-                    <Button
-                      type="button"
-                      variant="secondary"
-                      size="icon"
-                      className="h-9 w-9 shrink-0"
-                      onClick={() => browseDirectory(setAmazonDir)}
-                    >
-                      <FolderOpen className="h-4 w-4" />
-                    </Button>
-                  </div>
-                </div>
-                <Button
-                  variant="secondary"
-                  onClick={handleAmazonPosterScan}
-                  disabled={amazonScanning}
-                  className="w-full rounded-lg h-9 text-sm font-medium"
-                >
-                  {amazonScanning ? "正在扫描..." : "开始扫描"}
-                </Button>
-              </CardContent>
-            </Card>
-          </div>
-        )}
-
-        {/* ── 维护管理 ─────────────────────────── */}
-        {activeTab === "maintenance" && (
-          <div className="space-y-8 animate-in fade-in slide-in-from-bottom-4 duration-500">
-            <div className="px-1">
-              <h2 className="text-base font-semibold mb-1">维护与管理</h2>
-              <p className="text-muted-foreground text-xs">补全演员资料，保持资料库更完整</p>
-            </div>
-
-            <div className="grid gap-6 md:grid-cols-2">
-              <PersonToolCard
-                title="人物工具"
-                description="诊断连接状态并同步人物头像与简介"
-                headerExtra={
-                  <Select
-                    value={selectedPersonServer}
-                    onValueChange={(v) => setSelectedPersonServer(v as "jellyfin" | "emby")}
-                    disabled={anyPersonSyncRunning || anyPersonCheckPending}
-                  >
-                    <SelectTrigger className="h-9 w-[140px] rounded-lg">
-                      <SelectValue />
-                    </SelectTrigger>
-                    <SelectContent>
-                      <SelectItem value="jellyfin">Jellyfin</SelectItem>
-                      <SelectItem value="emby">Emby</SelectItem>
-                    </SelectContent>
-                  </Select>
-                }
-                {...personToolProps}
-                className="md:col-span-2"
-              />
-
-              <Card className="rounded-xl border shadow-sm md:col-span-2">
-                <CardHeader>
-                  <div className="flex items-center gap-2">
-                    <div className="p-1.5 bg-primary/8 rounded-lg">
-                      <Link2 className="h-5 w-5 text-primary" />
-                    </div>
-                    <div>
-                      <CardTitle className="text-sm font-medium">软链接管理</CardTitle>
-                      <CardDescription className="text-xs">在不同目录间建立文件组织结构映射</CardDescription>
-                    </div>
-                  </div>
-                </CardHeader>
-                <CardContent className="space-y-6">
-                  <div className="grid gap-6 md:grid-cols-2">
-                    <div className="grid gap-2">
-                      <Label htmlFor="sourceDir" className="text-xs font-medium text-muted-foreground">
-                        源目录
-                      </Label>
-                      <div className="flex gap-2">
-                        <Input
-                          id="sourceDir"
-                          value={sourceDir}
-                          onChange={(e) => setSourceDir(e.target.value)}
-                          className="h-9 bg-muted/30 rounded-lg border-none focus:ring-2 flex-1"
-                          placeholder="原始视频存放目录"
-                        />
-                        <Button
-                          type="button"
-                          variant="secondary"
-                          size="icon"
-                          className="h-9 w-9 shrink-0"
-                          onClick={() => browseDirectory(setSourceDir)}
-                        >
-                          <FolderOpen className="h-4 w-4" />
-                        </Button>
-                      </div>
-                    </div>
-                    <div className="grid gap-2">
-                      <Label htmlFor="destDir" className="text-xs font-medium text-muted-foreground">
-                        目标目录
-                      </Label>
-                      <div className="flex gap-2">
-                        <Input
-                          id="destDir"
-                          value={destDir}
-                          onChange={(e) => setDestDir(e.target.value)}
-                          className="h-9 bg-muted/30 rounded-lg border-none focus:ring-2 flex-1"
-                          placeholder="软链接存放目录"
-                        />
-                        <Button
-                          type="button"
-                          variant="secondary"
-                          size="icon"
-                          className="h-9 w-9 shrink-0"
-                          onClick={() => browseDirectory(setDestDir)}
-                        >
-                          <FolderOpen className="h-4 w-4" />
-                        </Button>
-                      </div>
-                    </div>
-                  </div>
-                  <div className="flex items-center space-x-3 p-3 bg-muted/20 rounded-xl">
-                    <Checkbox
-                      id="copyFiles"
-                      checked={copyFiles}
-                      onCheckedChange={(checked) => setCopyFiles(!!checked)}
-                    />
-                    <Label htmlFor="copyFiles" className="text-xs leading-tight cursor-pointer font-medium">
-                      同时同步 NFO, 图片及字幕等附属文件
-                    </Label>
-                  </div>
-                  <Button
-                    variant="secondary"
-                    onClick={handleCreateSymlink}
-                    disabled={createSymlinkMut.isPending}
-                    className="w-full rounded-lg h-9 text-sm font-medium"
-                  >
-                    {createSymlinkMut.isPending ? "正在处理..." : "立即建立映射"}
-                  </Button>
-                </CardContent>
-              </Card>
-            </div>
-          </div>
-        )}
-
-        {/* ── 垃圾清理 ─────────────────────────── */}
-        {activeTab === "cleanup" && (
-          <div className="space-y-8 animate-in fade-in slide-in-from-bottom-4 duration-500">
-            <div className="px-1">
-              <h2 className="text-base font-semibold mb-1">垃圾清理</h2>
-              <p className="text-muted-foreground text-xs">快速找出并清理无用的附带文件</p>
-            </div>
-
-            <Card className="rounded-xl border shadow-sm overflow-hidden">
-              <CardHeader>
-                <CardTitle className="text-sm font-medium">文件清理工具</CardTitle>
-                <CardDescription className="text-xs">选择目录和文件类型，先预览再删除</CardDescription>
-              </CardHeader>
-              <CardContent className="space-y-6">
-                <div className="flex flex-col lg:flex-row gap-4 items-end">
-                  <div className="grid gap-2 flex-1 w-full">
-                    <Label htmlFor="clean-path" className="text-xs font-medium text-muted-foreground">
-                      扫描目录
-                    </Label>
-                    <div className="flex gap-2">
-                      <Input
-                        id="clean-path"
-                        value={cleanPath}
-                        onChange={(e) => setCleanPath(e.target.value)}
-                        placeholder="/path/to/library"
-                        className="h-9 bg-muted/30 rounded-lg border-none focus:ring-2 flex-1"
-                      />
-                      <Button
-                        type="button"
-                        variant="secondary"
-                        size="icon"
-                        className="h-9 w-9 shrink-0"
-                        onClick={() => browseDirectory(setCleanPath)}
-                      >
-                        <FolderOpen className="h-4 w-4" />
-                      </Button>
-                    </div>
-                  </div>
-                  <Button
-                    variant="secondary"
-                    onClick={scanCleanupCandidates}
-                    disabled={cleanupScanning}
-                    className="h-9 px-8 rounded-lg shrink-0 text-sm font-medium"
-                  >
-                    {cleanupScanning ? "正在扫描..." : "开始扫描"}
-                  </Button>
-                </div>
-
-                <div className="space-y-4">
-                  <div className="flex items-center justify-between">
-                    <Label className="text-xs font-medium text-muted-foreground">文件类型过滤</Label>
-                    <div className="flex items-center space-x-2">
-                      <Checkbox
-                        id="include-subdirs"
-                        checked={includeSubdirs}
-                        onCheckedChange={(checked) => setIncludeSubdirs(!!checked)}
-                      />
-                      <Label htmlFor="include-subdirs" className="text-xs font-medium cursor-pointer">
-                        包含子目录
-                      </Label>
-                    </div>
-                  </div>
-
-                  <div className="flex flex-wrap gap-2">
-                    {CLEANUP_PRESET_EXTENSIONS.map((ext) => (
-                      <button
-                        key={ext}
-                        type="button"
-                        onClick={() => toggleCleanExtension(ext)}
-                        className={cn(
-                          "px-3 py-1.5 rounded-lg text-xs font-mono transition-all border",
-                          cleanExtensions.includes(ext)
-                            ? "bg-primary/10 border-primary/30 text-primary"
-                            : "bg-muted/20 border-transparent text-muted-foreground hover:bg-muted/40",
-                        )}
-                      >
-                        {ext}
-                      </button>
-                    ))}
-                  </div>
-
-                  <div className="flex gap-2 max-w-sm">
-                    <Input
-                      value={cleanCustomExt}
-                      onChange={(e) => setCleanCustomExt(e.target.value)}
-                      placeholder="自定义扩展名, 如 .bak"
-                      className="h-8 text-xs bg-muted/20 rounded-lg border-none"
-                    />
-                    <Button variant="ghost" size="sm" onClick={handleAddCustomExtension} className="h-8">
-                      添加
-                    </Button>
-                  </div>
-                </div>
-
-                {cleanupDeleting && (
-                  <div className="grid gap-2">
-                    <div className="flex justify-between text-xs text-muted-foreground font-medium">
-                      <span>正在删除文件...</span>
-                      <span>{cleanupProgress}%</span>
-                    </div>
-                    <Progress value={cleanupProgress} className="h-2 bg-muted/30" />
-                  </div>
-                )}
-
-                <div className="rounded-xl border bg-card overflow-hidden">
-                  <div className="overflow-x-auto">
-                    <table className="w-full text-xs">
-                      <thead>
-                        <tr className="bg-muted/30 text-muted-foreground">
-                          <th className="px-4 py-3 text-left font-semibold uppercase tracking-wider w-20">类型</th>
-                          <th className="px-4 py-3 text-left font-semibold uppercase tracking-wider">文件路径</th>
-                          <th className="px-4 py-3 text-left font-semibold uppercase tracking-wider w-24">大小</th>
-                          <th className="px-4 py-3 text-left font-semibold uppercase tracking-wider w-40">最后修改</th>
-                        </tr>
-                      </thead>
-                      <tbody className="divide-y divide-muted/20">
-                        {cleanupPreviewRows.length === 0 ? (
-                          <tr>
-                            <td colSpan={4} className="px-4 py-12 text-center text-muted-foreground italic">
-                              暂无待清理文件
-                            </td>
-                          </tr>
-                        ) : (
-                          cleanupPreviewRows.map((item) => (
-                            <tr key={item.path} className="hover:bg-muted/5 transition-colors">
-                              <td className="px-4 py-3 font-mono text-primary/70">{item.ext || "-"}</td>
-                              <td className="px-4 py-3 font-mono truncate max-w-md" title={item.path}>
-                                {item.path}
-                              </td>
-                              <td className="px-4 py-3 text-muted-foreground">
-                                {formatBytes(item.size, { fractionDigits: 2 })}
-                              </td>
-                              <td className="px-4 py-3 text-muted-foreground text-[10px]">
-                                {item.lastModified ?? "-"}
-                              </td>
-                            </tr>
-                          ))
-                        )}
-                      </tbody>
-                    </table>
-                  </div>
-                </div>
-
-                <div className="flex items-center justify-between p-4 bg-muted/20 rounded-xl">
-                  <div className="text-sm">
-                    <span className="text-muted-foreground">匹配文件: </span>
-                    <span className="font-bold">{cleanupCandidates.length}</span>
-                    <span className="mx-2 opacity-30">|</span>
-                    <span className="text-muted-foreground">总大小: </span>
-                    <span className="font-bold text-destructive">
-                      {formatBytes(cleanupTotalSize, { fractionDigits: 2 })}
-                    </span>
-                  </div>
-                  <Button
-                    variant="destructive"
-                    onClick={() => setCleanupConfirmOpen(true)}
-                    disabled={cleanupCandidates.length === 0 || cleanupDeleting}
-                    className="rounded-lg px-6 h-9 text-sm font-medium"
-                  >
-                    <Trash2 className="h-4 w-4 mr-2" />
-                    确认清理
-                  </Button>
-                </div>
-              </CardContent>
-            </Card>
-          </div>
-        )}
-
-        {/* ── 实用工具 ─────────────────────────── */}
-        {activeTab === "utility" && (
-          <div className="space-y-8 animate-in fade-in slide-in-from-bottom-4 duration-500">
-            <div className="px-1">
-              <h2 className="text-base font-semibold mb-1">实用工具</h2>
-              <p className="text-muted-foreground text-xs">提供批量翻译、缺番查找等轻量辅助功能</p>
-            </div>
-
-            <div className="grid gap-8 md:grid-cols-1 items-start">
-              <Card className="rounded-xl border shadow-sm">
-                <CardHeader>
-                  <div className="flex items-center gap-2">
-                    <div className="p-1.5 bg-primary/8 rounded-lg">
-                      <FileSearch className="h-5 w-5 text-primary" />
-                    </div>
-                    <div>
-                      <CardTitle className="text-sm font-medium">批量翻译 NFO</CardTitle>
-                      <CardDescription className="text-xs">
-                        扫描现有媒体库中的 NFO，使用当前 LLM 配置批量翻译标题和简介后回写
-                      </CardDescription>
-                    </div>
-                  </div>
-                </CardHeader>
-                <CardContent className="space-y-6">
-                  <div className="grid gap-2">
-                    <Label htmlFor="batch-translate-dir" className="text-xs font-medium text-muted-foreground">
-                      目标目录
-                    </Label>
-                    <div className="flex gap-2">
-                      <Input
-                        id="batch-translate-dir"
-                        value={batchTranslateDir}
-                        onChange={(e) => setBatchTranslateDir(e.target.value)}
-                        placeholder="输入已刮削完成的媒体目录"
-                        className="h-9 bg-muted/30 rounded-lg border-none focus:ring-2 flex-1"
-                      />
-                      <Button
-                        type="button"
-                        variant="secondary"
-                        size="icon"
-                        className="h-9 w-9 shrink-0"
-                        onClick={() => browseDirectory(setBatchTranslateDir)}
-                      >
-                        <FolderOpen className="h-4 w-4" />
-                      </Button>
-                    </div>
-                    <p className="text-[11px] leading-relaxed text-muted-foreground">
-                      该工具使用当前配置中的 LLM 模型、Base URL 与 API Key，独立于主刮削翻译流程。
-                    </p>
-                  </div>
-
-                  <div className="flex flex-col gap-3 sm:flex-row">
-                    <Button
-                      variant="secondary"
-                      onClick={handleBatchTranslateScan}
-                      disabled={batchTranslateScanning || batchTranslateApplying}
-                      className="flex-1 rounded-lg h-9 text-sm font-medium"
-                    >
-                      {batchTranslateScanning ? "正在扫描..." : "扫描待翻译条目"}
-                    </Button>
-                    <Button
-                      variant="secondary"
-                      onClick={handleBatchTranslateApply}
-                      disabled={batchTranslateApplying || batchTranslateScanning || batchTranslateItems.length === 0}
-                      className="flex-1 rounded-lg h-9 text-sm font-medium"
-                    >
-                      {batchTranslateApplying ? "正在批量翻译..." : "开始批量翻译并回写"}
-                    </Button>
-                  </div>
-
-                  <div className="flex flex-wrap items-center gap-3 text-xs">
-                    <span className="text-muted-foreground">待处理条目</span>
-                    <Badge variant="secondary">{batchTranslateItems.length}</Badge>
-                    <span className="text-muted-foreground">待处理字段</span>
-                    <Badge variant="secondary">{batchTranslatePendingFieldCount}</Badge>
-                    {batchTranslateResults.length > 0 && (
-                      <>
-                        <span className="text-muted-foreground">本次执行结果</span>
-                        <Badge variant="secondary">{batchTranslateResults.length}</Badge>
-                      </>
-                    )}
-                  </div>
-
-                  <div className="rounded-xl border bg-card overflow-hidden">
-                    <div className="overflow-x-auto">
-                      <table className="w-full text-xs">
-                        <thead>
-                          <tr className="bg-muted/30 text-muted-foreground">
-                            <th className="px-4 py-3 text-left font-semibold uppercase tracking-wider w-28">番号</th>
-                            <th className="px-4 py-3 text-left font-semibold uppercase tracking-wider">标题</th>
-                            <th className="px-4 py-3 text-left font-semibold uppercase tracking-wider w-40">
-                              待处理字段
-                            </th>
-                            <th className="px-4 py-3 text-left font-semibold uppercase tracking-wider w-72">NFO</th>
-                          </tr>
-                        </thead>
-                        <tbody className="divide-y divide-muted/20">
-                          {batchTranslatePreviewRows.length === 0 ? (
-                            <tr>
-                              <td colSpan={4} className="px-4 py-12 text-center text-muted-foreground italic">
-                                暂无待翻译条目
-                              </td>
-                            </tr>
-                          ) : (
-                            batchTranslatePreviewRows.map((item) => (
-                              <tr key={item.filePath} className="hover:bg-muted/5 transition-colors">
-                                <td className="px-4 py-3 font-mono font-medium">{item.number}</td>
-                                <td className="px-4 py-3">{item.title}</td>
-                                <td className="px-4 py-3">
-                                  <div className="flex flex-wrap gap-2">
-                                    {item.pendingFields.map((field) => (
-                                      <Badge key={`${item.filePath}-${field}`} variant="secondary">
-                                        {field === "title" ? "标题" : "简介"}
-                                      </Badge>
-                                    ))}
-                                  </div>
-                                </td>
-                                <td className="px-4 py-3 font-mono text-[11px] text-muted-foreground break-all">
-                                  {item.nfoPath}
-                                </td>
-                              </tr>
-                            ))
-                          )}
-                        </tbody>
-                      </table>
-                    </div>
-                  </div>
-
-                  {batchTranslateResults.length > 0 && (
-                    <div className="space-y-3">
-                      <div className="px-1">
-                        <Label className="text-xs font-medium text-muted-foreground">最近一次执行结果</Label>
-                      </div>
-                      <div className="rounded-xl border bg-card overflow-hidden">
-                        <div className="overflow-x-auto">
-                          <table className="w-full text-xs">
-                            <thead>
-                              <tr className="bg-muted/30 text-muted-foreground">
-                                <th className="px-4 py-3 text-left font-semibold uppercase tracking-wider w-24">
-                                  状态
-                                </th>
-                                <th className="px-4 py-3 text-left font-semibold uppercase tracking-wider w-28">
-                                  番号
-                                </th>
-                                <th className="px-4 py-3 text-left font-semibold uppercase tracking-wider w-36">
-                                  已写回字段
-                                </th>
-                                <th className="px-4 py-3 text-left font-semibold uppercase tracking-wider">结果</th>
-                              </tr>
-                            </thead>
-                            <tbody className="divide-y divide-muted/20">
-                              {batchTranslateResultRows.map((item) => {
-                                const partial = !item.success && item.translatedFields.length > 0;
-                                return (
-                                  <tr
-                                    key={`${item.filePath}-${item.nfoPath}`}
-                                    className="hover:bg-muted/5 transition-colors"
-                                  >
-                                    <td className="px-4 py-3">
-                                      <Badge variant="secondary">
-                                        {item.success ? "成功" : partial ? "部分成功" : "失败"}
-                                      </Badge>
-                                    </td>
-                                    <td className="px-4 py-3 font-mono font-medium">{item.number}</td>
-                                    <td className="px-4 py-3">
-                                      <div className="flex flex-wrap gap-2">
-                                        {item.translatedFields.length === 0 ? (
-                                          <span className="text-muted-foreground">-</span>
-                                        ) : (
-                                          item.translatedFields.map((field) => (
-                                            <Badge key={`${item.nfoPath}-${field}`} variant="secondary">
-                                              {field === "title" ? "标题" : "简介"}
-                                            </Badge>
-                                          ))
-                                        )}
-                                      </div>
-                                    </td>
-                                    <td className="px-4 py-3">
-                                      <div className="space-y-1">
-                                        {item.savedNfoPath && (
-                                          <div className="font-mono text-[11px] text-muted-foreground break-all">
-                                            {item.savedNfoPath}
-                                          </div>
-                                        )}
-                                        {item.error && <div className="text-destructive">{item.error}</div>}
-                                      </div>
-                                    </td>
-                                  </tr>
-                                );
-                              })}
-                            </tbody>
-                          </table>
-                        </div>
-                      </div>
-                    </div>
-                  )}
-                </CardContent>
-              </Card>
-
-              {/* 缺番查找工具 */}
-              <Card className="rounded-xl border shadow-sm">
-                <CardHeader>
-                  <div className="flex items-center gap-2">
-                    <div className="p-1.5 bg-primary/8 rounded-lg">
-                      <Link2 className="h-5 w-5 text-primary" />
-                    </div>
-                    <div>
-                      <CardTitle className="text-sm font-medium">缺番查找</CardTitle>
-                      <CardDescription className="text-xs">根据编号范围和现有列表快速识别缺失的番号</CardDescription>
-                    </div>
-                  </div>
-                </CardHeader>
-                <CardContent className="space-y-6">
-                  <div className="grid gap-4 md:grid-cols-2">
-                    <div className="grid gap-2">
-                      <Label htmlFor="missing-prefix" className="text-xs font-medium text-muted-foreground">
-                        番号前缀
-                      </Label>
-                      <Input
-                        id="missing-prefix"
-                        placeholder="例如: ABC"
-                        value={missingPrefix}
-                        onChange={(e) => setMissingPrefix(e.target.value)}
-                        className="h-9 bg-muted/30 rounded-lg border-none focus:ring-2"
-                      />
-                    </div>
-                    <div className="grid gap-2">
-                      <Label htmlFor="missing-range" className="text-xs font-medium text-muted-foreground">
-                        数字范围
-                      </Label>
-                      <Input
-                        id="missing-range"
-                        placeholder="例如: 1-120"
-                        value={missingRange}
-                        onChange={(e) => setMissingRange(e.target.value)}
-                        className="h-9 bg-muted/30 rounded-lg border-none focus:ring-2"
-                      />
-                    </div>
-                  </div>
-
-                  <div className="grid gap-2">
-                    <Label htmlFor="existing-numbers" className="text-xs font-medium text-muted-foreground">
-                      已存在番号列表
-                    </Label>
-                    <Textarea
-                      id="existing-numbers"
-                      value={existingNumbers}
-                      onChange={(e) => setExistingNumbers(e.target.value)}
-                      placeholder="ABC-001, ABC-002, ABC-004..."
-                      className="min-h-[120px] bg-muted/30 rounded-lg border-none focus:ring-2 p-4 text-sm"
-                    />
-                  </div>
-
-                  <Button
-                    variant="secondary"
-                    onClick={handleFindMissing}
-                    className="w-full rounded-lg h-9 text-sm font-medium"
-                  >
-                    开始查找缺失番号
-                  </Button>
-
-                  <div className="space-y-3">
-                    <div className="flex items-center justify-between px-1">
-                      <Label className="text-xs font-medium text-muted-foreground">查找结果</Label>
-                      {missingSummary && <span className="text-[10px] font-medium text-primary">{missingSummary}</span>}
-                    </div>
-                    <div className="rounded-xl border bg-muted/10 overflow-hidden">
-                      <table className="w-full text-xs">
-                        <thead>
-                          <tr className="bg-muted/30 text-muted-foreground">
-                            <th className="px-4 py-2 text-left w-16">序号</th>
-                            <th className="px-4 py-2 text-left">建议番号</th>
-                          </tr>
-                        </thead>
-                        <tbody>
-                          {missingPreviewRows.length === 0 ? (
-                            <tr>
-                              <td colSpan={2} className="px-4 py-8 text-center text-muted-foreground italic">
-                                无查找结果
-                              </td>
-                            </tr>
-                          ) : (
-                            missingPreviewRows.map((row) => (
-                              <tr
-                                key={row.number}
-                                className="border-t border-muted/10 hover:bg-muted/5 transition-colors"
-                              >
-                                <td className="px-4 py-2 text-muted-foreground">{row.index}</td>
-                                <td className="px-4 py-2 font-mono font-medium">{row.number}</td>
-                              </tr>
-                            ))
-                          )}
-                        </tbody>
-                      </table>
-                    </div>
-                  </div>
-                </CardContent>
-              </Card>
-            </div>
-          </div>
-        )}
-
-        {/* Bottom spacing */}
-        <div className="h-10" />
-      </div>
-
-      <AmazonPosterDialog
-        open={amazonPosterDialogOpen}
-        onOpenChange={setAmazonPosterDialogOpen}
-        items={amazonPosterScanItems}
-      />
-
-      <Dialog open={cleanupConfirmOpen} onOpenChange={setCleanupConfirmOpen}>
-        <DialogContent className="rounded-xl">
-          <DialogHeader>
-            <DialogTitle>确认清理文件</DialogTitle>
-            <DialogDescription>
-              将永久删除 <span className="font-bold text-foreground">{cleanupCandidates.length}</span> 个文件 (约{" "}
-              <span className="font-bold text-destructive">{formatBytes(cleanupTotalSize, { fractionDigits: 2 })}</span>
-              )。此操作无法撤销。
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter className="gap-2">
+    <div ref={pageScrollRef} className="h-full w-full overflow-y-auto bg-surface-canvas scroll-smooth">
+      {selectedToolId === null ? (
+        <main className="mx-auto w-full max-w-[1120px] px-6 py-8 md:px-8 lg:px-10 lg:py-10">
+          <ToolOverview onSelect={handleSelectTool} />
+        </main>
+      ) : (
+        <main className="mx-auto flex w-full max-w-[1120px] flex-col px-6 py-6 md:px-8 lg:px-10 lg:py-8">
+          <div className="sticky top-0 z-10 mb-6 w-fit rounded-full bg-surface-canvas/92 pt-1 backdrop-blur-sm">
             <Button
-              variant="ghost"
-              onClick={() => setCleanupConfirmOpen(false)}
-              disabled={cleanupDeleting}
-              className="rounded-lg"
+              type="button"
+              variant="secondary"
+              size="icon"
+              onClick={handleBackToOverview}
+              className="h-12 w-12 rounded-full bg-surface-low text-foreground shadow-[0_12px_24px_rgba(15,23,42,0.06)] hover:bg-surface-raised/80"
             >
-              取消
+              <ArrowLeft className="h-5 w-5" />
             </Button>
-            <Button
-              variant="destructive"
-              onClick={handleDeleteCleanupCandidates}
-              disabled={cleanupDeleting}
-              className="rounded-lg px-8"
-            >
-              {cleanupDeleting ? "正在清理..." : "确认删除"}
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+          </div>
+
+          <ToolDetail toolId={selectedToolId} />
+          <div className="h-2" />
+        </main>
+      )}
     </div>
   );
-}
-
-function shouldKeepForCleanup(item: FileItem, extensionSet: Set<string>) {
-  if (item.type !== "file") return false;
-  const ext = extensionFromName(item.name);
-  return ext.length > 0 && extensionSet.has(ext);
 }
