+++ src/renderer/src/store/uiStore.ts
@@ -7,7 +7,6 @@ interface UIState {
   showPreviewPanel: boolean;
   workbenchMode: "scrape" | "maintenance";
   settingsActiveTab: string;
-  toolActiveTab: string;
 
   setSelectedResultId: (id: string | null) => void;
   setSidebarOpen: (open: boolean) => void;
@@ -16,7 +15,6 @@ interface UIState {
   setShowPreviewPanel: (show: boolean) => void;
   setWorkbenchMode: (mode: "scrape" | "maintenance") => void;
   setSettingsActiveTab: (tab: string) => void;
-  setToolActiveTab: (tab: string) => void;
 }
 
 export const useUIStore = create<UIState>((set) => ({
@@ -26,7 +24,6 @@ export const useUIStore = create<UIState>((set) => ({
   showPreviewPanel: true,
   workbenchMode: "scrape",
   settingsActiveTab: "",
-  toolActiveTab: "",
 
   setSelectedResultId: (id) => set({ selectedResultId: id }),
   setSidebarOpen: (open) => set({ sidebarOpen: open }),
@@ -35,5 +32,4 @@ export const useUIStore = create<UIState>((set) => ({
   setShowPreviewPanel: (show) => set({ showPreviewPanel: show }),
   setWorkbenchMode: (mode) => set({ workbenchMode: mode }),
   setSettingsActiveTab: (tab) => set({ settingsActiveTab: tab }),
-  setToolActiveTab: (tab) => set({ toolActiveTab: tab }),
 }));
